+++ lib/dashlogin.dart
@@ -1,0 +1,28 @@
+import 'package:flutter/material.dart';
+
+
+
+class Dashlogin extends StatelessWidget {
+
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text("DashBoard"),
+        actions: [
+          IconButton(onPressed: (){}, icon: Icon(Icons.logout))
+        ],
+      ),
+      body: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        crossAxisAlignment: CrossAxisAlignment.center,
+        children: [
+
+          Text("Cogrates you are logged in"),
+
+        ],
+      ),
+    );
+  }
+}
